+++ chia/data_layer/data_layer_server.py
@@ -13,11 +13,11 @@ import click
 from aiohttp import web
 
 from chia.data_layer.download_data import is_filename_valid
+from chia.server.signal_handlers import SignalHandlers
 from chia.server.upnp import UPnP
 from chia.util.chia_logging import initialize_logging
 from chia.util.config import load_config
 from chia.util.default_root import DEFAULT_ROOT_PATH
-from chia.util.misc import SignalHandlers
 from chia.util.network import WebServer
 from chia.util.path import path_from_root
 from chia.util.setproctitle import setproctitle
